requirements.txt
asgiref==3.4.1
autopep8==1.5.7
cffi==1.14.6
cryptography==3.4.7
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
